+++ testdata/app/decorators/admin/report_decorator.rb
@@ -1,0 +1,9 @@
+module Admin
+  class ReportDecorator < ApplicationDecorator
+    delegate_all
+
+    def summary_line
+      "#{object.title} (#{object.status})"
+    end
+  end
+end